requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
plotly==5.17.0
scikit-learn==1.3.2
python-dotenv==1.0.0
schedule==1.2.0
openpyxl==3.1.2
seaborn==0.12.2
matplotlib==3.8.2
